priv/python/pyproject.toml
[tool.poetry]
name = "lux-python"
version = "0.1.0"
description = "Python components for the Lux framework"
authors = ["Robert Parcus <robertparcus@spectrallabs.xyz>"]
readme = "README.md"
packages = [{include = "lux"}]

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^6.15.1"
eth-tester = {extras = ["py-evm"], version = "^0.9.0b1"}
nltk = "^3.9.1"
setuptools = "^68.0.0"
erlport = "^0.6"
hyperliquid-python-sdk = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=lux --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 